requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# requirements.txt — CourtAccess AI runtime dependencies
#
# NOTE: The canonical source of truth is pyproject.toml.
# Docker images install via:  pip install -e .  (API)
#                              pip install -e '.[airflow]'  (Airflow workers)
#
# This file is a convenience for:
#   - Local venv setup without pyproject.toml editable install
#   - CI environments that prefer a flat requirements file
#   - Teammates using:  pip install -r requirements.txt
#
# To regenerate from pyproject.toml:
#   uv pip compile pyproject.toml -o requirements.txt
# ══════════════════════════════════════════════════════════════════════════════

# ── Web API ──────────────────────────────────────────────────────────────────
fastapi[standard]>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# ── Cloud ────────────────────────────────────────────────────────────────────
google-cloud-storage>=2.16.0
google-cloud-logging>=3.10.0

# ── AI / ML (ARM64-compatible) ───────────────────────────────────────────────
faster-whisper>=1.0.0
ctranslate2>=4.0.0
transformers>=4.40.0
groq>=0.9.0

# ── PDF / OCR ─────────────────────────────────────────────────────────────────
pymupdf>=1.23.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# ── Web scraping ─────────────────────────────────────────────────────────────
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
playwright>=1.40.0

# ── Data versioning ──────────────────────────────────────────────────────────
dvc>=3.50.0

# ── Utilities ────────────────────────────────────────────────────────────────
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# ── Airflow (Airflow containers only) ────────────────────────────────────────
# apache-airflow>=3.1.7   # Uncomment if installing into a non-Airflow base image
